+++ api/src/main/kotlin/io/github/llh4github/lotus/api/ApiApp.kt
@@ -8,7 +8,7 @@ import org.springframework.boot.runApplication
  * Created At 2023/12/29 22:49
  * @author llh
  */
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = ["io.github.llh4github.lotus.*"])
 class ApiApp
