+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/MenuResourceService.kt
@@ -1,6 +1,6 @@
 package io.github.llh4github.lotus.api.service.auth
 
-import io.github.llh4github.lotus.api.service.BaseService
+import io.github.llh4github.lotus.model.BaseService
 import io.github.llh4github.lotus.model.auth.MenuResource
 import io.github.llh4github.lotus.model.auth.dto.MenuResourceAddInput
 import io.github.llh4github.lotus.model.auth.dto.MenuResourceUpdateInput
