+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/PurviewCodeService.kt
@@ -1,6 +1,6 @@
 package io.github.llh4github.lotus.api.service.auth
 
-import io.github.llh4github.lotus.api.service.BaseService
+import io.github.llh4github.lotus.model.BaseService
 import io.github.llh4github.lotus.model.auth.PurviewCode
 import io.github.llh4github.lotus.model.auth.dto.PurviewCodeAddInput
 import io.github.llh4github.lotus.model.auth.dto.PurviewCodeUpdateInput
