+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/UrlResourceService.kt
@@ -1,6 +1,6 @@
 package io.github.llh4github.lotus.api.service.auth
 
-import io.github.llh4github.lotus.api.service.BaseService
+import io.github.llh4github.lotus.model.BaseService
 import io.github.llh4github.lotus.model.HttpMethodEnums
 import io.github.llh4github.lotus.model.auth.UrlResource
 import io.github.llh4github.lotus.model.auth.dto.UrlResourceAddInput
