+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/UserService.kt
@@ -1,6 +1,6 @@
 package io.github.llh4github.lotus.api.service.auth
 
-import io.github.llh4github.lotus.api.service.BaseService
+import io.github.llh4github.lotus.model.BaseService
 import io.github.llh4github.lotus.model.auth.User
 import io.github.llh4github.lotus.model.auth.dto.UserAddInput
 import io.github.llh4github.lotus.model.auth.dto.UserUpdateInput
