+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/impl/RoleServiceImplImpl.kt
@@ -1,10 +1,10 @@
 package io.github.llh4github.lotus.api.service.auth.impl
 
-import io.github.llh4github.lotus.api.dao.RoleDao
 import io.github.llh4github.lotus.api.exceptions.auth.RoleException
-import io.github.llh4github.lotus.api.service.BaseServiceImpl
 import io.github.llh4github.lotus.api.service.auth.RoleService
+import io.github.llh4github.lotus.model.BaseServiceImpl
 import io.github.llh4github.lotus.model.auth.Role
+import io.github.llh4github.lotus.model.auth.RoleDao
 import io.github.llh4github.lotus.model.auth.dto.RoleAddInput
 import io.github.llh4github.lotus.model.auth.dto.RoleUpdateInput
 import io.github.oshai.kotlinlogging.KotlinLogging
