+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/impl/UrlResourceServiceImpl.kt
@@ -1,16 +1,12 @@
 package io.github.llh4github.lotus.api.service.auth.impl
 
-import io.github.llh4github.lotus.api.dao.UrlResourceDao
 import io.github.llh4github.lotus.api.exceptions.auth.UrlResourceException
-import io.github.llh4github.lotus.api.service.BaseServiceImpl
 import io.github.llh4github.lotus.api.service.auth.UrlResourceService
+import io.github.llh4github.lotus.model.BaseServiceImpl
 import io.github.llh4github.lotus.model.HttpMethodEnums
-import io.github.llh4github.lotus.model.auth.UrlResource
+import io.github.llh4github.lotus.model.auth.*
 import io.github.llh4github.lotus.model.auth.dto.UrlResourceAddInput
 import io.github.llh4github.lotus.model.auth.dto.UrlResourceUpdateInput
-import io.github.llh4github.lotus.model.auth.id
-import io.github.llh4github.lotus.model.auth.method
-import io.github.llh4github.lotus.model.auth.path
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.babyfish.jimmer.sql.kt.ast.expression.eq
 import org.babyfish.jimmer.sql.kt.ast.expression.ne
