+++ api/src/main/kotlin/io/github/llh4github/lotus/api/service/auth/impl/UserServiceImpl.kt
@@ -1,10 +1,10 @@
 package io.github.llh4github.lotus.api.service.auth.impl
 
-import io.github.llh4github.lotus.api.dao.UserDao
 import io.github.llh4github.lotus.api.exceptions.auth.UserException
-import io.github.llh4github.lotus.api.service.BaseServiceImpl
 import io.github.llh4github.lotus.api.service.auth.UserService
+import io.github.llh4github.lotus.model.BaseServiceImpl
 import io.github.llh4github.lotus.model.auth.User
+import io.github.llh4github.lotus.model.auth.UserDao
 import io.github.llh4github.lotus.model.auth.dto.UserAddInput
 import io.github.llh4github.lotus.model.auth.dto.UserUpdateInput
 import io.github.oshai.kotlinlogging.KotlinLogging
