+++ model/src/main/kotlin/io/github/llh4github/lotus/model/auth/MenuResourceDao.kt
@@ -1,13 +1,14 @@
-package io.github.llh4github.lotus.api.dao
+package io.github.llh4github.lotus.model.auth
 
-import io.github.llh4github.lotus.model.auth.MenuResource
+
+import io.github.llh4github.lotus.model.BaseDao
 import org.springframework.stereotype.Component
 import kotlin.reflect.KClass
 
 /**
  *
  *
- * Created At 2024/1/23 18:23
+ * Created At 2024/1/27 19:47
  * @author llh
  */
 @Component
