+++ model/src/main/kotlin/io/github/llh4github/lotus/model/auth/UserDao.kt
@@ -1,20 +1,22 @@
-package io.github.llh4github.lotus.api.dao
+package io.github.llh4github.lotus.model.auth
 
-import io.github.llh4github.lotus.model.auth.User
-import io.github.llh4github.lotus.model.auth.username
+
+import io.github.llh4github.lotus.model.BaseDao
 import org.babyfish.jimmer.View
 import org.babyfish.jimmer.sql.fetcher.Fetcher
 import org.babyfish.jimmer.sql.kt.ast.expression.eq
 import org.springframework.stereotype.Component
-import org.springframework.stereotype.Repository
 import kotlin.reflect.KClass
 
+/**
+ *
+ *
+ * Created At 2024/1/27 19:47
+ * @author llh
+ */
 @Component
 class UserDao : BaseDao<User>() {
-    override val entityType: KClass<User>
-        get() = User::class
-
-
+    override val entityType = User::class
     fun findByUsername(username: String, fetcher: Fetcher<User>? = null): User? {
         return createQuery {
             where(table.username eq username)
